adding infra to separated folder

+++ infra/src/infra/application/use-cases/admin/delete-admin.ts
@@ -1,12 +1,17 @@
 import { Message } from "../../../../responses/response";
-import { Admin, AdminProps } from "../../../entities/domain/admin";
 import { AdminRepository } from "../../../repository/admin-repository";
 
 export class DeleteAdmin {
   constructor(private adminRepository: AdminRepository) {}
 
   async execute(adminId: string): Promise<Message> {
-    const response = await this.adminRepository.deleteAdmin(adminId);
-    return response;
+    const checkAdmin = await this.adminRepository.checkAdmin(adminId);
+
+    if (checkAdmin.statusCode === 200) {
+      const response = await this.adminRepository.deleteAdmin(adminId);
+      return response;
+    }
+
+    return checkAdmin;
   }
 }
